+++ cli/commands/cluster/health.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/coreos/etcd/etcdserver/etcdserverpb"
 	"github.com/sensu/sensu-go/cli"
 	"github.com/sensu/sensu-go/cli/commands/helpers"
 	"github.com/sensu/sensu-go/cli/elements/table"
@@ -22,13 +23,29 @@ func HealthCommand(cli *cli.SensuCli) *cobra.Command {
 			if err != nil {
 				return err
 			}
-			return helpers.Print(cmd, cli.Config.Format(), printToTable, nil, result)
+			clusterHealth := result.ClusterHealth
+			alarms := result.Alarms
+			err = helpers.Print(cmd, cli.Config.Format(), printHealthToTable, nil, clusterHealth)
+			if err != nil {
+				return err
+			}
+
+			if alarms != nil {
+				err = helpers.Print(cmd, cli.Config.Format(), printAlarmsToTable, nil, alarms)
+				if err != nil {
+					return err
+				}
+			}
+
+			return nil
 		},
 	}
+
+	helpers.AddFormatFlag(cmd.Flags())
 	return cmd
 }
 
-func printToTable(result interface{}, w io.Writer) {
+func printHealthToTable(result interface{}, w io.Writer) {
 	table := table.New([]*table.Column{
 		{
 			Title:       "ID",
@@ -76,5 +93,33 @@ func printToTable(result interface{}, w io.Writer) {
 		},
 	})
 
+	table.Render(w, result)
+}
+
+func printAlarmsToTable(result interface{}, w io.Writer) {
+	table := table.New([]*table.Column{
+		{
+			Title:       "ID",
+			ColumnStyle: table.PrimaryTextStyle,
+			CellTransformer: func(data interface{}) string {
+				alarm, ok := data.(*etcdserverpb.AlarmMember)
+				if !ok {
+					return cli.TypeError
+				}
+				return fmt.Sprintf("%x", alarm.GetMemberID())
+			},
+		},
+		{
+			Title:       "Alarm Type",
+			ColumnStyle: table.PrimaryTextStyle,
+			CellTransformer: func(data interface{}) string {
+				alarm, ok := data.(*etcdserverpb.AlarmMember)
+				if !ok {
+					return cli.TypeError
+				}
+				return alarm.Alarm.String()
+			},
+		},
+	})
 	table.Render(w, result)
 }